Narrow the syntax node end position to 32 bits
The end position sat in a usize beside the twelve-byte source location, so
every syntax node carried a four-byte hole and an eight-byte field where four
bytes reach as far as the location itself does. The field is now u32 and fills
the hole. The accessor still answers in usize, so no call site changed.

The AST arena for the openmpi configure file falls from 230839768 to 225218808
bytes.

+++ src/Expressions.cpp
@@ -48,7 +48,7 @@ pure fn Expression::source_end_position() const wontthrow -> usize
 
 fn Expression::set_source_end_position(usize position) wontthrow -> void
 {
-  m_source_end_position = position;
+  m_source_end_position = static_cast<u32>(position);
 }
 
 cold fn Expression::to_ast_string(usize layer) const throws -> String
